Move all source code o src folder and rename Tests to tests
