+++ src/Command/AnnotatedCommandTrait.php
