+++ src/Command/ConfiguredCommandTrait.php
