+++ src/Entity/ScheduledTaskLog.php
