+++ src/Utils/HashHelper.php
