+++ src/Utils/StringHelper.php
