+++ tests/Fixtures/FooBundle/Command/AnnotatedCommand.php
@@ -1,19 +1,24 @@
 <?php
 
-namespace Tests\Fixtures\FooBundle\Command;
+namespace Goksagun\SchedulerBundle\Tests\Fixtures\FooBundle\Command;
 
+use Goksagun\SchedulerBundle\Annotation\Schedule;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 
-class DatabasedCommand extends Command
+/**
+ * @Schedule(name="schedule:annotate", expression="*\/10 * * * *")
+ * @Schedule(name="schedule:annotate --foo=bar", expression="* * * * *")
+ */
+class AnnotatedCommand extends Command
 {
     protected function configure()
     {
         $this
-            ->setName('schedule:database')
-            ->setDescription('This command scheduled from database.')
+            ->setName('schedule:annotate')
+            ->setDescription('This command do not return any output.')
             ->addOption('foo', null, InputOption::VALUE_OPTIONAL);
     }
 
@@ -25,7 +30,7 @@ class DatabasedCommand extends Command
             $output->writeln("This is an foo: {$foo}");
         }
 
-        $output->writeln("Hello from schedule by database");
+        $output->writeln("Hello from schedule by annotation");
 
         return 0;
     }
