+++ tests/Fixtures/FooBundle/Command/DatabasedCommand.php
@@ -1,24 +1,19 @@
 <?php
 
-namespace Tests\Fixtures\FooBundle\Command;
+namespace Goksagun\SchedulerBundle\Tests\Fixtures\FooBundle\Command;
 
-use Goksagun\SchedulerBundle\Annotation\Schedule;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 
-/**
- * @Schedule(name="schedule:annotate", expression="*\/10 * * * *")
- * @Schedule(name="schedule:annotate --foo=bar", expression="* * * * *")
- */
-class AnnotatedCommand extends Command
+class DatabasedCommand extends Command
 {
     protected function configure()
     {
         $this
-            ->setName('schedule:annotate')
-            ->setDescription('This command do not return any output.')
+            ->setName('schedule:database')
+            ->setDescription('This command scheduled from database.')
             ->addOption('foo', null, InputOption::VALUE_OPTIONAL);
     }
 
@@ -30,7 +25,7 @@ class AnnotatedCommand extends Command
             $output->writeln("This is an foo: {$foo}");
         }
 
-        $output->writeln("Hello from schedule by annotation");
+        $output->writeln("Hello from schedule by database");
 
         return 0;
     }
