+++ tests/Utils/ArrayHelperTest.php
